Create ViewModels and move Dtos
