+++ CollabClothing.Appication/CollabClothing.Appication.csproj
@@ -7,6 +7,7 @@
   <ItemGroup>
     <ProjectReference Include="..\CollabClothing.Utilities\CollabClothing.Utilities.csproj" />
     <ProjectReference Include="..\CollabClothing.WebApp\CollabClothing.WebApp.csproj" />
+    <ProjectReference Include="..\CollabClothing.ViewModels\CollabClothing.ViewModels.csproj" />
   </ItemGroup>
 
   <ItemGroup>
