+++ CollabClothing.ViewModels/Catalog/Products/Manage/GetRequestPagingProduct.cs
@@ -1,11 +1,11 @@
-﻿using CollabClothing.Appication.Dtos;
+﻿using CollabClothing.ViewModels.Common;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace CollabClothing.Appication.Catalog.Products.Dtos.Manage
+namespace CollabClothing.ViewModels.Catalog.Products.Manage
 {
     public class GetRequestPagingProduct : PagingRequestBase
     {
