+++ CollabClothing.ViewModels/Catalog/Products/Manage/ProductCreateRequest.cs
@@ -1,12 +1,14 @@
-﻿using System;
+﻿using CollabClothing.WebApp.Models;
+using Microsoft.AspNetCore.Http;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace CollabClothing.Appication.Catalog.Products.Dtos
+namespace CollabClothing.ViewModels.Catalog.Products.Manage
 {
-    public class ProductViewModel
+    public class ProductCreateRequest
     {
         public string Id { get; set; }
         public string ProductName { get; set; }
@@ -19,5 +21,9 @@ namespace CollabClothing.Appication.Catalog.Products.Dtos
         public string Description { get; set; }
         public string Slug { get; set; }
         public string Details { get; set; }
+
+        public Category Category { get; set; }
+        public Size Size { get; set; }
+
     }
 }
