+++ CollabClothing.ViewModels/Catalog/Products/Public/GetRequestPagingProduct.cs
@@ -1,4 +1,4 @@
-﻿using CollabClothing.Appication.Dtos;
+﻿using CollabClothing.ViewModels.Common;
 using System;
 using System.Collections.Generic;
 using System.Linq;
