+++ CollabClothing.ViewModels/CollabClothing.ViewModels.csproj
@@ -1,0 +1,13 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net5.0</TargetFramework>
+  </PropertyGroup>
+ <ItemGroup>
+    <ProjectReference Include="..\CollabClothing.WebApp\CollabClothing.WebApp.csproj" />
+  </ItemGroup>
+  <ItemGroup>
+    <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="5.0.14" />
+  </ItemGroup>
+
+</Project>
