+++ CollabClothing.ViewModels/Common/PageResult.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace CollabClothing.Appication.Dtos
+namespace CollabClothing.ViewModels.Common
 {
     public class PageResult<T>
     {
